rust: WireType::Sequence decoder + refactoring
Adds initial support for decoding sequences, using a simple state
machine parser similar to the message parser.

This commit also includes a large amount of refactoring to split apart
the decoder into separate modules for each type.
